+++ src/Compute/Compute/ChangeLog.md
@@ -34,6 +34,11 @@
     - `New-AzVM`
     - `New-AzVmssConfig`
     - `New-AzVMSS`
+* Added cmdlets to create, update, delete, and get new Azure resource: Ssh Public Key
+    - `New-AzSshKey`
+    - `Remove-AzSshKey`
+    - `Get-AzSshKey`
+    - `Update-AzSshKey`
 
 ## Version 4.12.0
 * Updated the `Set-AzVMDiskEncryptionExtension` cmdlet to support ADE extension migration from two pass (version with AAD input parameters) to single pass (version without AAD input parameters).